combined io block and tree
