+++ src/amaranth/hdlTree.py
@@ -2,18 +2,8 @@ from amaranth import *
 from amaranth import Elaboratable, Module, Signal, unsigned, Const 
 from amaranth.sim import Simulator
 from amaranth.back import rtlil, verilog
-# import numpy as np
-from amaranth.cli import main
+import numpy as np
 
-
-def generate_verilog(root, filename="tree.v"):
-    m = Module()
-    s_input = Signal(unsigned(25))
-    s_output = Signal(unsigned(8))
-    root.rec_set_m_i_o(m, s_input, s_output)
-    root.activate()
-    with open(filename, "w") as f:
-        f.write(verilog.convert(m, ports=[s_input, s_output]))
 
 class node:
     def __init__(self, m, input_, output_, index=-1, class_=-1,  child_yes=None, child_no=None, isLeaf=False):
@@ -27,7 +17,6 @@ class node:
         self.isLeaf = isLeaf
         self.child_yes = child_yes
         self.child_no = child_no
-
 
     def activate(self):
         if not self.isLeaf:
@@ -60,17 +49,17 @@ class node:
 def test_Tree(x, y, root, test_name="test"):
     # set correct input output signals
     m = Module()
-    s_input = Signal(unsigned(25))
+    s_image = Signal(unsigned(25))
     s_output = Signal(unsigned(8))
     # set the input output signals
-    root.rec_set_m_i_o(m, s_input, s_output)
+    root.rec_set_m_i_o(m, s_image, s_output)
     root.activate()
     # convert x from booleqn array to int
     x_ = 0
     for i, bit in enumerate(x):
         x_ += bit*2**i
     # init the input
-    m.d.sync+= s_input.eq(Const(int(x_)))
+    m.d.sync+= s_image.eq(Const(int(x_)))
     async def test_bench(ctx):
         global res
         for _ in range(2):
@@ -83,3 +72,134 @@ def test_Tree(x, y, root, test_name="test"):
 
     with sim.write_vcd(test_name+".vcd"):
         sim.run()
+
+
+class IO_Block:
+    def __init__(self, m:Module, In_raw : Signal, In_image : Signal):
+
+        # In_raw: system input, 8 bit bus, 
+        #         In_raw [7:5]: indicate which row of the image the input belong to
+        #         In_raw [4:0]: The value of each pixel on a row of a 5X5 image
+        self.In_raw = In_raw
+
+        # In_image: 25 bit bus representing the 5x5 image
+        #           [4:0] -> 1st row, ... [24:20] -> last row 
+        self.In_image = In_image
+
+        self.m = m
+
+    def activate(self):
+        # create signal for In_raw[7:5]
+        with self.m.Switch(self.In_raw[0:3]):
+            with self.m.Case("000"):
+                self.m.d.sync += self.In_image[0:5].eq(self.In_raw[3:8])
+            with self.m.Case("001"):
+                self.m.d.sync += self.In_image[5:10].eq(self.In_raw[3:8])
+            with self.m.Case("010"):
+                self.m.d.sync += self.In_image[10:15].eq(self.In_raw[3:8])
+            with self.m.Case("011"):
+                self.m.d.sync += self.In_image[15:20].eq(self.In_raw[3:8])
+            with self.m.Case("100"):
+                self.m.d.sync += self.In_image[20:25].eq(self.In_raw[3:8])
+
+def test_IO_Block(x, test_name="test"):
+    # create a module
+    m = Module()
+    s_input = Signal(unsigned(8))
+    s_output = Signal(unsigned(25))
+    block = IO_Block(m, s_input, s_output)
+    block.activate()
+
+    #fomat the input
+    shape_data = np.array(x).reshape(5, -1).tolist()
+    data =[]
+    for i,d in enumerate(shape_data):
+        x_ = 0
+        for j, bit in enumerate(d):
+            x_ += bit*2**j
+        data.append((i,x_))
+
+    async def test_bench(ctx):
+        for p in data:
+            # set the input
+            ctx.set(s_input, p[1]*2**3 + p[0])
+            await ctx.tick()
+        final = f"{ctx.get(s_output):b}"
+        res = []
+        for i in final:
+            res.append(bool(int(i)))
+        while len(res) < 25:
+            res.insert(0, False)
+        res = np.array(res[::-1])
+        assert np.all(res == np.array(x))
+
+    sim = Simulator(m)
+    sim.add_clock(1e-6)
+    sim.add_testbench(test_bench)
+    with sim.write_vcd(test_name+".vcd"):
+        sim.run()
+    
+
+def test_combined(x,y, root:node, test_name="test"):
+    # create a module
+    m = Module()
+    s_input = Signal(unsigned(8))
+    s_image = Signal(unsigned(25))
+    s_output = Signal(unsigned(8))
+    rst = Signal()
+    io = IO_Block(m, s_input, s_image)
+    io.activate()
+    root.rec_set_m_i_o(m, s_image, s_output)
+    root.activate()
+    m = ResetInserter({'sync':rst})(m)
+    
+    #fomat the input
+    shape_data = np.array(x).reshape(5, -1).tolist()
+    data =[]
+    for i,d in enumerate(shape_data):
+        x_ = 0
+        for j, bit in enumerate(d):
+            x_ += bit*2**j
+        data.append((i,x_))
+
+    async def test_bench(ctx):
+        for p in data:
+            # set the input
+            ctx.set(s_input, p[1]*2**3 + p[0])
+            await ctx.tick()
+        assert ctx.get(s_output) == y
+        ctx.set(s_input, 0)
+        ctx.set(rst, 1)
+        await ctx.tick()
+        ctx.set(rst, 0)
+        await ctx.tick()
+        assert ctx.get(s_image) == 0
+
+    sim = Simulator(m)
+    sim.add_clock(1e-6)
+    sim.add_testbench(test_bench)
+    with sim.write_vcd(test_name+".vcd"):
+        sim.run()
+
+
+def generate_verilog(root:node, filename="tree.v"):
+    # create the module and key signals
+    m = Module()
+    s_input = Signal(unsigned(8))
+    s_image = Signal(unsigned(25))
+    s_output = Signal(unsigned(8))
+    rst = Signal()
+    # create the IO block and the tree
+    io = IO_Block(m, s_input, s_image)
+    io.activate()
+    # set the tree signals
+    root.rec_set_m_i_o(m, s_image, s_output)
+    root.activate()
+    # add reset
+    m = ResetInserter({'sync':rst})(m)
+    # convert the module to verilog
+    with open(filename, "w") as f:
+        f.write(verilog.convert(m, ports=[s_input, s_output]))
+
+    print("Verilog file generated")
+    print(f"File: {filename}")
